@@ -1,13 +1,11 @@
 <template>
   <span>
-    <span
-      class="playlist__prev"
-      @click="prev">prev</span>
-    <span class="playlist__counter"> page {{ pagecurr }} of {{ maxvalue }} </span>
-    <span
-      class="playlist__next"
-      @click="next">next</span>
-    <br>
+    <span class="playlist__prev" @click="prev">prev</span>
+    <span class="playlist__counter">
+      page {{ pagecurr }} of {{ maxvalue }}
+    </span>
+    <span class="playlist__next" @click="next">next</span>
+    <br />
   </span>
 </template>
 <script>
@@ -15,24 +13,24 @@ export default {
   props: {
     pagecurr: {
       type: Number,
-      default: 0
+      default: 0,
     },
     maxvalue: {
       type: Number,
-      default: 5
-    }
+      default: 5,
+    },
   },
   methods: {
-    next () {
+    next() {
       if (this.pagecurr < this.maxvalue) {
-        this.$emit('update', this.pagecurr + 1)
+        this.$emit("update", this.pagecurr + 1);
       }
     },
-    prev () {
+    prev() {
       if (this.pagecurr > 1) {
-        this.$emit('update', this.pagecurr - 1)
+        this.$emit("update", this.pagecurr - 1);
       }
-    }
-  }
-}
+    },
+  },
+};
 </script>
