@@ -1,11 +1,7 @@
 <template>
   <div class="playlist__player">
-    <audio
-      controls
-      class="audio">
-      <source
-        :src="file"
-        type="audio/mpeg">
+    <audio controls class="audio">
+      <source :src="file" type="audio/mpeg" />
       Тег audio не поддерживается вашим браузером.
       <a :href="file">Скачайте музыку</a>.
     </audio>
@@ -17,13 +13,13 @@ export default {
   props: {
     file: {
       type: String,
-      default: 'none'
-    }
+      default: "none",
+    },
   },
   computed: {
-    shortName () {
-      return this.file.substring(this.file.length - 17)
-    }
-  }
-}
+    shortName() {
+      return this.file.substring(this.file.length - 17);
+    },
+  },
+};
 </script>
