@@ -1,26 +1,34 @@
 /* eslint-disable no-console */
 
-import { register } from 'register-service-worker'
+import { register } from "register-service-worker";
 
-if (process.env.NODE_ENV === 'production') {
+if (process.env.NODE_ENV === "production") {
   register(`${process.env.BASE_URL}service-worker.js`, {
-    ready () {
+    ready() {
       console.log(
-        'App is being served from cache by a service worker.\n' +
-        'For more details, visit https://goo.gl/AFskqB'
-      )
+        "App is being served from cache by a service worker.\n" +
+          "For more details, visit https://goo.gl/AFskqB"
+      );
     },
-    cached () {
-      console.log('Content has been cached for offline use.')
+    registered() {
+      console.log("Service worker has been registered.");
     },
-    updated () {
-      console.log('New content is available; please refresh.')
+    cached() {
+      console.log("Content has been cached for offline use.");
     },
-    offline () {
-      console.log('No internet connection found. App is running in offline mode.')
+    updatefound() {
+      console.log("New content is downloading.");
     },
-    error (error) {
-      console.error('Error during service worker registration:', error)
+    updated() {
+      console.log("New content is available; please refresh.");
     },
-  })
+    offline() {
+      console.log(
+        "No internet connection found. App is running in offline mode."
+      );
+    },
+    error(error) {
+      console.error("Error during service worker registration:", error);
+    },
+  });
 }
