@@ -1,90 +1,87 @@
 <!-- Use preprocessors via the lang attribute! e.g. <template lang="pug"> -->
 <template>
   <div id="app">
-    <div
-      v-if="files.length"
-      class="playlist">
-      <complayer :file="fullurl(id)"/>
+    <div v-if="files.length" class="playlist">
+      <complayer :file="fullurl(id)" />
       <comtrack
-        v-for="(f,k) in pagearr"
+        v-for="(f, k) in pagearr"
         :key="k"
         :index="f"
         :pagecurr="pagecurr"
         :selectid="id + 1"
-        @select="selectfile"/>
-      <compagination
-        :pagecurr="pagecurr"
-        :maxvalue="5"
-        @update="updtpage"
+        @select="selectfile"
       />
+      <compagination :pagecurr="pagecurr" :maxvalue="5" @update="updtpage" />
     </div>
   </div>
 </template>
 
 <script>
-import complayer from '@/components/complayer.vue'
-import comtrack from '@/components/comtrack.vue'
-import compagination from '@/components/compagination.vue'
+import complayer from "@/components/complayer.vue";
+import comtrack from "@/components/comtrack.vue";
+import compagination from "@/components/compagination.vue";
 
 export default {
   components: {
     complayer,
     comtrack,
-    compagination
+    compagination,
   },
-  data () {
+  data() {
     return {
-      filelink: '',
+      filelink: "",
       files: [],
       id: 0,
       folder: 1,
       pagearr: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
       pagecurr: 1,
-      status: 'none',
-      baseurl: 'https://ivanshavliuga.github.io/simples/audio/booksshop/telepat/'
-    }
+      status: "none",
+      baseurl:
+        "https://ivanshavliuga.github.io/simples/audio/booksshop/telepat/",
+    };
   },
-  created () {
+  created() {
     for (let i = 1; i <= 10; i++) {
-      this.files.push(i + '_tm.MP3')
+      this.files.push(i + "_tm.MP3");
     }
   },
   methods: {
-    fullurl (id) {
-      const rid = this.getID(id)
-      if (this.pagecurr < 2 && rid < 9) return this.baseurl + 'telepatia000' + (rid + 1) + '.mp3'
-      else return this.baseurl + 'telepatia00' + (rid + 1) + '.mp3'
+    fullurl(id) {
+      const rid = this.getID(id);
+      if (this.pagecurr < 2 && rid < 9)
+        return this.baseurl + "telepatia000" + (rid + 1) + ".mp3";
+      else return this.baseurl + "telepatia00" + (rid + 1) + ".mp3";
     },
     /* fullurl (id) {
       const ext = (this.folder < 2) ? ('_tm.MP3') : ('_tm.mp3')
       const rid = this.getID(id)
       return this.baseurl + this.folder + '/' + (rid + 1) + ext
     }, */
-    playevent () {
-      this.status = 'play'
+    playevent() {
+      this.status = "play";
     },
-    selectfile (id) {
-      let audiotag = document.querySelector('audio')
-      audiotag.src = this.fullurl(id)
-      this.id = id
-      audiotag.play()
+    selectfile(id) {
+      let audiotag = document.querySelector("audio");
+      audiotag.src = this.fullurl(id);
+      this.id = id;
+      audiotag.play();
     },
-    getID (x) {
-      return (this.pagecurr - 1) * 10 + x
+    getID(x) {
+      return (this.pagecurr - 1) * 10 + x;
     },
-    updtpage (pg) {
-      this.pagecurr = pg
+    updtpage(pg) {
+      this.pagecurr = pg;
     },
-    updtfolder (fld) {
-      this.folder = fld
-    }
-  }
-}
+    updtfolder(fld) {
+      this.folder = fld;
+    },
+  },
+};
 </script>
 
 <!-- Use preprocessors via the lang attribute! e.g. <style lang="scss"> -->
 <style>
-  /* img:before {
+/* img:before {
   content: "?";
   position: absolute;
   background-color: #343423;
@@ -114,9 +111,9 @@ img:after {
   background-color: @bodybackground!important;
 } */
 
-  body{
-    background: #606060;
-  }
+body {
+  background: #606060;
+}
 .playlist {
   width: 550px;
   padding: 0;
@@ -125,46 +122,48 @@ img:after {
   box-shadow: 1px 1px 3px 5px #aaa;
 }
 .playlist__item {
-    background-color: #fcfcfc;
-    position: relative;
-    padding-left: 15px;
-  }
-  .playlist__number, .playlist__name {
-    display: inline;
-    font-size: 20px;
-    font-weight: normal;
-    line-height: 36px;
-    margin: 0 10px;
-  }
-  .playlist__player {
-    border-bottom: 7px double #999;
-    padding: 15px 0;
-    margin: 15px 0;
-    background-color: #dfdfdf;
-  }
-  .playlist__number {
-    color: red;
-  }
-  .playlist__panel {
-    margin: 15px 25px;
-  }
-  .playlist__prev, .playlist__next {
-    color: red;
-    cursor: pointer;
-    padding: 10px 5px;
-    line-height: 40px;
-    font-size: 16px;
-  }
-  .playlist__debug {
-    color: grey;
-    cursor: pointer;
-    padding: 10px 5px;
-    line-height: 30px;
-    margin-left: 20px;
-  }
-  .playlist__counter {
-    font-size: 21px;
-  }
+  background-color: #fcfcfc;
+  position: relative;
+  padding-left: 15px;
+}
+.playlist__number,
+.playlist__name {
+  display: inline;
+  font-size: 20px;
+  font-weight: normal;
+  line-height: 36px;
+  margin: 0 10px;
+}
+.playlist__player {
+  border-bottom: 7px double #999;
+  padding: 15px 0;
+  margin: 15px 0;
+  background-color: #dfdfdf;
+}
+.playlist__number {
+  color: red;
+}
+.playlist__panel {
+  margin: 15px 25px;
+}
+.playlist__prev,
+.playlist__next {
+  color: red;
+  cursor: pointer;
+  padding: 10px 5px;
+  line-height: 40px;
+  font-size: 16px;
+}
+.playlist__debug {
+  color: grey;
+  cursor: pointer;
+  padding: 10px 5px;
+  line-height: 30px;
+  margin-left: 20px;
+}
+.playlist__counter {
+  font-size: 21px;
+}
 audio {
   display: block;
   width: 500px;
@@ -173,7 +172,7 @@ audio {
   margin: 2px auto;
   background-color: #dfdfdf;
 }
-  .itemactive {
-    background-color: #42a912;
-  }
+.itemactive {
+  background-color: #42a912;
+}
 </style>
