@@ -1,8 +1,8 @@
 module.exports = {
   env: {
-    mocha: true
+    mocha: true,
   },
   rules: {
-    'import/no-extraneous-dependencies': 'off'
-  }
-}
+    "import/no-extraneous-dependencies": "off",
+  },
+};
