@@ -1,4 +1,3 @@
-import { expect } from "chai";
 import { shallowMount } from "@vue/test-utils";
 import HelloWorld from "@/components/HelloWorld.vue";
 
@@ -6,8 +5,8 @@ describe("HelloWorld.vue", () => {
   it("renders props.msg when passed", () => {
     const msg = "new message";
     const wrapper = shallowMount(HelloWorld, {
-      propsData: { msg },
+      props: { msg },
     });
-    expect(wrapper.text()).to.include(msg);
+    expect(wrapper.text()).toMatch(msg);
   });
 });
